add dash support.  turn off default Interactive mode.
dash is used on some platforms as the default system shell (i.e. it
masquerades as /bin/sh).  dash will suspend itself if it is put into
interactive mode (via the -i option) and is not connected to a
terminal.  because of this, the default is to *not* put the shell in
interactive mode.  If interactive mode is required, use the Expect mode.